Move model code to SudokuLogic
